Abstracting away common methods that will be used in WeekView
